hice un pequeño cambio en la base de datos, especificamente en la tabla Detalle_pedido

+++ db_practicasprof.sql
@@ -54,7 +54,7 @@ CREATE TABLE Detalle_pedido (
     id_detalle_pedido INT AUTO_INCREMENT PRIMARY KEY,
     id_pedido INT NOT NULL,
     id_producto INT NOT NULL,
-    cantidad INT NOT NULL,
+    cantidad DECIMAL(10, 2) NOT NULL,
     FOREIGN KEY (id_pedido) REFERENCES Pedido(id_pedido),
     FOREIGN KEY (id_producto) REFERENCES Producto(id_producto)
 ) ENGINE=InnoDB;
